requirements.txt
Flask==2.1.0
Pillow==9.1.0
Werkzeug==2.2.2
flask-wtf==1.0.1
python-dotenv==0.20.0 